updates: minor UI changes

+++ server/src/main/java/com/chirp/server/controllers/PostController.java
@@ -52,7 +52,7 @@ public class PostController {
 	}
 
 	@GetMapping("/list/count/{userId}")
-	public int getUserPostCount(@PathVariable String userId) {
+	public long getUserPostCount(@PathVariable String userId) {
 		try {
 			return postService.getPostCountPerUser(userId);
 		} catch (Exception e) {
